+++ exporters/otlp/include/opentelemetry/exporters/otlp/otlp_http_log_record_exporter_runtime_options.h
@@ -5,6 +5,7 @@
 
 #include <memory>
 
+#include "opentelemetry/exporters/otlp/otlp_json_writer_factory.h"
 #include "opentelemetry/sdk/common/thread_instrumentation.h"
 #include "opentelemetry/version.h"
 
@@ -23,6 +24,8 @@ struct OPENTELEMETRY_EXPORT OtlpHttpLogRecordExporterRuntimeOptions
 
   std::shared_ptr<sdk::common::ThreadInstrumentation> thread_instrumentation =
       std::shared_ptr<sdk::common::ThreadInstrumentation>(nullptr);
+
+  std::shared_ptr<JsonWriterFactory> json_writer_factory;
 };
 
 }  // namespace otlp