+++ exporters/otlp/src/otlp_http_log_record_exporter.cc
@@ -21,7 +21,6 @@
 #include "opentelemetry/nostd/string_view.h"
 #include "opentelemetry/sdk/common/exporter_utils.h"
 #include "opentelemetry/sdk/common/global_log_handler.h"
-#include "opentelemetry/sdk/common/thread_instrumentation.h"
 #include "opentelemetry/version.h"
 
 // clang-format off
@@ -45,6 +44,33 @@ namespace exporter
 namespace otlp
 {
 
+namespace
+{
+
+OtlpHttpClientOptions MakeClientOptions(
+    const OtlpHttpLogRecordExporterOptions &options,
+    const OtlpHttpLogRecordExporterRuntimeOptions &runtime_options)
+{
+  OtlpHttpClientOptions client_options(
+      options.url, options.ssl_insecure_skip_verify, options.ssl_ca_cert_path,
+      options.ssl_ca_cert_string, options.ssl_client_key_path, options.ssl_client_key_string,
+      options.ssl_client_cert_path, options.ssl_client_cert_string, options.ssl_min_tls,
+      options.ssl_max_tls, options.ssl_cipher, options.ssl_cipher_suite, options.content_type,
+      options.json_bytes_mapping, options.compression, options.use_json_name, options.console_debug,
+      options.timeout, options.http_headers, options.retry_policy_max_attempts,
+      options.retry_policy_initial_backoff, options.retry_policy_max_backoff,
+      options.retry_policy_backoff_multiplier, runtime_options.thread_instrumentation
+#ifdef ENABLE_ASYNC_EXPORT
+      ,
+      options.max_concurrent_requests, options.max_requests_per_connection
+#endif
+  );
+  client_options.json_writer_factory = runtime_options.json_writer_factory;
+  return client_options;
+}
+
+}  // namespace
+
 OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter()
     : OtlpHttpLogRecordExporter(OtlpHttpLogRecordExporterOptions())
 {}
@@ -53,37 +79,7 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
     const OtlpHttpLogRecordExporterOptions &options)
     : options_(options),
       runtime_options_(),
-      http_client_(std::make_unique<OtlpHttpClient>(OtlpHttpClientOptions(
-          options.url,
-          options.ssl_insecure_skip_verify,
-          options.ssl_ca_cert_path,
-          options.ssl_ca_cert_string,
-          options.ssl_client_key_path,
-          options.ssl_client_key_string,
-          options.ssl_client_cert_path,
-          options.ssl_client_cert_string,
-          options.ssl_min_tls,
-          options.ssl_max_tls,
-          options.ssl_cipher,
-          options.ssl_cipher_suite,
-          options.content_type,
-          options.json_bytes_mapping,
-          options.compression,
-          options.use_json_name,
-          options.console_debug,
-          options.timeout,
-          options.http_headers,
-          options.retry_policy_max_attempts,
-          options.retry_policy_initial_backoff,
-          options.retry_policy_max_backoff,
-          options.retry_policy_backoff_multiplier,
-          std::shared_ptr<sdk::common::ThreadInstrumentation>{nullptr}
-#ifdef ENABLE_ASYNC_EXPORT
-          ,
-          options.max_concurrent_requests,
-          options.max_requests_per_connection
-#endif
-          )))
+      http_client_(std::make_unique<OtlpHttpClient>(MakeClientOptions(options, runtime_options_)))
 {}
 
 OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
@@ -91,37 +87,7 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
     const OtlpHttpLogRecordExporterRuntimeOptions &runtime_options)
     : options_(options),
       runtime_options_(runtime_options),
-      http_client_(std::make_unique<OtlpHttpClient>(OtlpHttpClientOptions(
-          options.url,
-          options.ssl_insecure_skip_verify,
-          options.ssl_ca_cert_path,
-          options.ssl_ca_cert_string,
-          options.ssl_client_key_path,
-          options.ssl_client_key_string,
-          options.ssl_client_cert_path,
-          options.ssl_client_cert_string,
-          options.ssl_min_tls,
-          options.ssl_max_tls,
-          options.ssl_cipher,
-          options.ssl_cipher_suite,
-          options.content_type,
-          options.json_bytes_mapping,
-          options.compression,
-          options.use_json_name,
-          options.console_debug,
-          options.timeout,
-          options.http_headers,
-          options.retry_policy_max_attempts,
-          options.retry_policy_initial_backoff,
-          options.retry_policy_max_backoff,
-          options.retry_policy_backoff_multiplier,
-          runtime_options.thread_instrumentation
-#ifdef ENABLE_ASYNC_EXPORT
-          ,
-          options.max_concurrent_requests,
-          options.max_requests_per_connection
-#endif
-          )))
+      http_client_(std::make_unique<OtlpHttpClient>(MakeClientOptions(options, runtime_options_)))
 {}
 
 OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
@@ -129,38 +95,8 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
     const std::shared_ptr<ext::http::client::HttpClientFactory> &factory)
     : options_(options),
       runtime_options_(),
-      http_client_(std::make_unique<OtlpHttpClient>(
-          OtlpHttpClientOptions(options.url,
-                                options.ssl_insecure_skip_verify,
-                                options.ssl_ca_cert_path,
-                                options.ssl_ca_cert_string,
-                                options.ssl_client_key_path,
-                                options.ssl_client_key_string,
-                                options.ssl_client_cert_path,
-                                options.ssl_client_cert_string,
-                                options.ssl_min_tls,
-                                options.ssl_max_tls,
-                                options.ssl_cipher,
-                                options.ssl_cipher_suite,
-                                options.content_type,
-                                options.json_bytes_mapping,
-                                options.compression,
-                                options.use_json_name,
-                                options.console_debug,
-                                options.timeout,
-                                options.http_headers,
-                                options.retry_policy_max_attempts,
-                                options.retry_policy_initial_backoff,
-                                options.retry_policy_max_backoff,
-                                options.retry_policy_backoff_multiplier,
-                                std::shared_ptr<sdk::common::ThreadInstrumentation>{nullptr}
-#ifdef ENABLE_ASYNC_EXPORT
-                                ,
-                                options.max_concurrent_requests,
-                                options.max_requests_per_connection
-#endif
-                                ),
-          factory))
+      http_client_(
+          std::make_unique<OtlpHttpClient>(MakeClientOptions(options, runtime_options_), factory))
 {}
 
 OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
@@ -169,38 +105,8 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
     const std::shared_ptr<ext::http::client::HttpClientFactory> &factory)
     : options_(options),
       runtime_options_(runtime_options),
-      http_client_(std::make_unique<OtlpHttpClient>(
-          OtlpHttpClientOptions(options.url,
-                                options.ssl_insecure_skip_verify,
-                                options.ssl_ca_cert_path,
-                                options.ssl_ca_cert_string,
-                                options.ssl_client_key_path,
-                                options.ssl_client_key_string,
-                                options.ssl_client_cert_path,
-                                options.ssl_client_cert_string,
-                                options.ssl_min_tls,
-                                options.ssl_max_tls,
-                                options.ssl_cipher,
-                                options.ssl_cipher_suite,
-                                options.content_type,
-                                options.json_bytes_mapping,
-                                options.compression,
-                                options.use_json_name,
-                                options.console_debug,
-                                options.timeout,
-                                options.http_headers,
-                                options.retry_policy_max_attempts,
-                                options.retry_policy_initial_backoff,
-                                options.retry_policy_max_backoff,
-                                options.retry_policy_backoff_multiplier,
-                                runtime_options.thread_instrumentation
-#ifdef ENABLE_ASYNC_EXPORT
-                                ,
-                                options.max_concurrent_requests,
-                                options.max_requests_per_connection
-#endif
-                                ),
-          factory))
+      http_client_(
+          std::make_unique<OtlpHttpClient>(MakeClientOptions(options, runtime_options_), factory))
 {}
 
 OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
@@ -208,38 +114,8 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
     std::shared_ptr<ext::http::client::HttpClient> http_client)
     : options_(options),
       runtime_options_(),
-      http_client_(std::make_unique<OtlpHttpClient>(
-          OtlpHttpClientOptions(options.url,
-                                options.ssl_insecure_skip_verify,
-                                options.ssl_ca_cert_path,
-                                options.ssl_ca_cert_string,
-                                options.ssl_client_key_path,
-                                options.ssl_client_key_string,
-                                options.ssl_client_cert_path,
-                                options.ssl_client_cert_string,
-                                options.ssl_min_tls,
-                                options.ssl_max_tls,
-                                options.ssl_cipher,
-                                options.ssl_cipher_suite,
-                                options.content_type,
-                                options.json_bytes_mapping,
-                                options.compression,
-                                options.use_json_name,
-                                options.console_debug,
-                                options.timeout,
-                                options.http_headers,
-                                options.retry_policy_max_attempts,
-                                options.retry_policy_initial_backoff,
-                                options.retry_policy_max_backoff,
-                                options.retry_policy_backoff_multiplier,
-                                std::shared_ptr<sdk::common::ThreadInstrumentation>{nullptr}
-#ifdef ENABLE_ASYNC_EXPORT
-                                ,
-                                options.max_concurrent_requests,
-                                options.max_requests_per_connection
-#endif
-                                ),
-          std::move(http_client)))
+      http_client_(std::make_unique<OtlpHttpClient>(MakeClientOptions(options, runtime_options_),
+                                                    std::move(http_client)))
 {}
 
 OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
@@ -248,38 +124,8 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
     std::shared_ptr<ext::http::client::HttpClient> http_client)
     : options_(options),
       runtime_options_(runtime_options),
-      http_client_(std::make_unique<OtlpHttpClient>(
-          OtlpHttpClientOptions(options.url,
-                                options.ssl_insecure_skip_verify,
-                                options.ssl_ca_cert_path,
-                                options.ssl_ca_cert_string,
-                                options.ssl_client_key_path,
-                                options.ssl_client_key_string,
-                                options.ssl_client_cert_path,
-                                options.ssl_client_cert_string,
-                                options.ssl_min_tls,
-                                options.ssl_max_tls,
-                                options.ssl_cipher,
-                                options.ssl_cipher_suite,
-                                options.content_type,
-                                options.json_bytes_mapping,
-                                options.compression,
-                                options.use_json_name,
-                                options.console_debug,
-                                options.timeout,
-                                options.http_headers,
-                                options.retry_policy_max_attempts,
-                                options.retry_policy_initial_backoff,
-                                options.retry_policy_max_backoff,
-                                options.retry_policy_backoff_multiplier,
-                                runtime_options.thread_instrumentation
-#ifdef ENABLE_ASYNC_EXPORT
-                                ,
-                                options.max_concurrent_requests,
-                                options.max_requests_per_connection
-#endif
-                                ),
-          std::move(http_client)))
+      http_client_(std::make_unique<OtlpHttpClient>(MakeClientOptions(options, runtime_options_),
+                                                    std::move(http_client)))
 {}
 
 OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(std::unique_ptr<OtlpHttpClient> http_client)